Resolves issue 20, duplicate blocked users, for real!
Changed the way BlockManager fetches the blocked user list. Before, it was passed in by the BuddyPanel. This was a problem because the BuddyPanel holds all friends, even blocked.

Now, the BlockManager fetches an alphabetical list of unblocked users directly from the model. This method was created in the model, getUnblockedFriendList(), for this purpose.

Thanks for the info, Jihoon.