Trying to modularize our MainController. My plan is to have each AccountData hold an interface type... which would be implemented by XMPPManager, TwitterManager, and ICQManager. Then, whenever we want to add a friend, for example, the controller would: 1. find the account to add to 2. call model.addFriend(account) 3. the AccountData will automatically connect with the server via the XMPPManager class, which holds the XMPPConnection smack field.
The power of this is that model.addFriend(account) will work for any connection, because the AccountData has a GenericConnection field, which could be any protocol.

This way, the MainController doesn't have any reference to a specific protocol. If an error occurs, BadConnectionException will be thrown, which is a general exception class.

Let me know your thoughts.