We now have a contact search. Enjoy. To use, just enter in characters at the bottom of the buddy panel in the search box.
Currently, it searches the nickname, userID, and status. It is not case sensitive. The algorithm is in UserData.

We can easily change the behaviour. For instance, we can make it cull friends instead of sorting them. Also, we can make a pop-up appear with search results upon pressing the search button.