Twitter poller is now functional!
TwitterManager now instantiates a thread that forces the MainController to update its server data every 30 seconds (changeable in TwitterManager). Currently, the MainController only checks existing (server) friends for status updates. Check the code out in MainController.refreshFriends().

Note: I am very unhappy with how it is currently implemented. Basically, now the MainController instantiates a new TwitterManager, which handles all connections with Twitter. This is fine. However, the TwitterManager then creates a thread that tells the MainController to check the TwitterManager for new updates. Very circular... The GoogleTalk/JabberManagers suffer the same sort of problem with their event listeners. I wonder if we can fix this logic?