YES, THREADED PICTURE UPDATES!!!!
Now the buddy list loads quickly, and the pictures slowly get added. This new thread is located in BuddyPanel... it is called PictureUpdateThread.


Known issue: if you enter a search, it takes a long, long, time for all the pictures to update. THis is because a new list is grabbed at each key stroke. update() has a similar, but harder to see, issue.